register default way of getting browser session
Clients of this lib should provide browser session mechanism
default is to just start a local binary of firefox in tests
