+++ lib/domkey.rb
@@ -1,5 +1,5 @@
 require 'domkey/version'
-require 'watir-webdriver'
+require 'domkey/browser'
 require 'domkey/view'
 require 'domkey/view/component'
 require 'domkey/view/component_collection'
@@ -14,12 +14,22 @@ module Domkey
   # current browser for testing session
   def self.browser
     return @browser if (@browser && @browser.exist?)
-    # simple browser
-    @browser = Watir::Browser.new
+    @browser = Browser.new
   end
 
   # sets current browser for testing session
   def self.browser=(b)
     @browser = b
+  end
+
+  # close browser on exit to cleanup after yourself
+  def self.close_browser_on_exit
+    return if @_close_browser_on_exit
+    at_exit {
+      if @browser && @browser.exists?
+        @browser.close
+      end
+    }
+    @_close_browser_on_exit = true
   end
 end
